feat: Added FullDetailCard story and renamed index to FullDetailCard
